+++ noodles-gtf/src/record/attributes.rs
@@ -67,16 +67,67 @@ fn parse_attributes(mut src: &str) -> io::Result<IndexMap<&str, Value<'_>>> {
 
     while !src.is_empty() {
         let (key, raw_value) = parse_field(&mut src)?;
+        let decoded_raw_value = escape_decode(raw_value)?;
 
         match map.entry(key) {
             Entry::Vacant(entry) => {
-                entry.insert(Value::String(raw_value));
+                entry.insert(Value::String(decoded_raw_value));
             }
-            Entry::Occupied(mut entry) => entry.get_mut().push(raw_value),
+            Entry::Occupied(mut entry) => entry.get_mut().push(decoded_raw_value),
         }
     }
 
     Ok(map)
+}
+
+fn escape_decode(s: &str) -> io::Result<Cow<'_, str>> {
+    const BACKSLASH: char = '\\';
+
+    if s.contains(BACKSLASH) {
+        unescape_string(s).map(Cow::from)
+    } else {
+        Ok(Cow::from(s))
+    }
+}
+
+fn unescape_string(s: &str) -> io::Result<String> {
+    const BACKSLASH: char = '\\';
+    const QUOTATION_MARK: char = '"';
+
+    enum State {
+        Normal,
+        Escape,
+    }
+
+    let mut dst = String::with_capacity(s.len());
+    let mut state = State::Normal;
+
+    for c in s.chars() {
+        match state {
+            State::Normal => {
+                if c == BACKSLASH {
+                    state = State::Escape;
+                } else {
+                    dst.push(c);
+                }
+            }
+            State::Escape => {
+                match c {
+                    BACKSLASH | QUOTATION_MARK => dst.push(c),
+                    _ => {
+                        return Err(io::Error::new(
+                            io::ErrorKind::InvalidData,
+                            "invalid escape sequence",
+                        ))
+                    }
+                }
+
+                state = State::Normal;
+            }
+        }
+    }
+
+    Ok(dst)
 }
 
 #[cfg(test)]
@@ -95,10 +146,13 @@ mod tests {
     fn test_get() -> io::Result<()> {
         let attributes = Attributes::try_new(r#"id 0; name "ndls";"#)?;
 
-        assert_eq!(attributes.get("id").transpose()?, Some(&Value::String("0")));
+        assert_eq!(
+            attributes.get("id").transpose()?,
+            Some(&Value::String(Cow::from("0")))
+        );
         assert_eq!(
             attributes.get("name").transpose()?,
-            Some(&Value::String("ndls"))
+            Some(&Value::String(Cow::from("ndls")))
         );
 
         assert!(attributes.get("comment").transpose()?.is_none());
@@ -116,29 +170,44 @@ mod tests {
         }
 
         f("", &[])?;
-        f("id 0;", &[("id", &Value::String("0"))])?;
-        f("id 0", &[("id", &Value::String("0"))])?;
+        f("id 0;", &[("id", &Value::String(Cow::from("0")))])?;
+        f("id 0", &[("id", &Value::String(Cow::from("0")))])?;
         f(
             r#"id 0; name "ndls";"#,
             &[
-                ("id", &Value::String("0")),
-                ("name", &Value::String("ndls")),
+                ("id", &Value::String(Cow::from("0"))),
+                ("name", &Value::String(Cow::from("ndls"))),
             ],
         )?;
         f(
             r#"id 0;name "ndls";"#,
             &[
-                ("id", &Value::String("0")),
-                ("name", &Value::String("ndls")),
+                ("id", &Value::String(Cow::from("0"))),
+                ("name", &Value::String(Cow::from("ndls"))),
             ],
         )?;
         f(
             r#"id 0;  name "ndls";  "#,
             &[
-                ("id", &Value::String("0")),
-                ("name", &Value::String("ndls")),
+                ("id", &Value::String(Cow::from("0"))),
+                ("name", &Value::String(Cow::from("ndls"))),
             ],
         )?;
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_escape_decode() -> io::Result<()> {
+        assert_eq!(escape_decode("")?, "");
+        assert_eq!(escape_decode("ndls")?, "ndls");
+        assert_eq!(escape_decode(r"nd\\ls")?, r"nd\ls");
+        assert_eq!(escape_decode(r#"nd\"ls\""#)?, r#"nd"ls""#);
+
+        assert!(matches!(
+            escape_decode(r#"nd\ls"#),
+            Err(e) if e.kind() == io::ErrorKind::InvalidData
+        ));
 
         Ok(())
     }